made correction to object deletion test; new api views

+++ models/location.py
@@ -3,7 +3,7 @@
 contains the Location class definition
 """
 from models.base_model import Base, BaseModel, Column
-from sqlalchemy import DECIMAL
+from sqlalchemy import Double
 from sqlalchemy.orm import relationship
 
 
@@ -13,8 +13,8 @@ class Location(BaseModel, Base):
     """
     __tablename__ = 'locations'
 
-    latitude = Column('latitude', DECIMAL(10, 6))
-    longitude = Column('longitude', DECIMAL(10, 6))
+    latitude = Column('latitude', Double)
+    longitude = Column('longitude', Double)
     # alternatively use coordinates - POINT '2D SPACIAL INDEX' datatype
 
     user = relationship('User', back_populates='location', uselist=False)
